refactor: move Logo to Header
