+++ src/components/Header/components/Logo/Logo.scss
